refactor: Move email components into subfolder
