+++ src/templates/EditorialNewsletter/email/Blockquote.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { fontFamilies } from '../../theme/fonts'
+import { fontFamilies } from '../../../theme/fonts'
 
 const blockquoteStyle = {
   marginBottom: '30px'
