+++ src/templates/EditorialNewsletter/email/Cover.js
