+++ src/templates/EditorialNewsletter/email/Header.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import colors from '../../theme/colors'
+import colors from '../../../theme/colors'
 
 export default () => {
   return (
