+++ src/templates/EditorialNewsletter/email/List.js
